--- instagram_to_discord/util.py
@@ -1,6 +1,0 @@
-def is_int(s):
-    try:
-        int(s)
-    except:
-        return False
-    return True
